DartBuddyREADME
To navigate through the pages use the top app bar (right click on mouse, pull from on touch screen devices). 
The player page and the main page have separate share functions.
To change players for a game go to player page select (or create) a player and tap the player one or two button.
The how to page explains the rules of the game and how to use the app.
The dartboard will also zoom on touchscreen devices.